Updated the test cases with more
